test: scope fixtures
Add scope to column-type ad database fixtures to avoid excessive
repeating code.
Merge database and connected_database fixture as there is no need for
both and the ambiguity of whether the database is connected is
problematic.
Modify fixtures that break database connection to immediately return it.
Add dropped_database fixture.